Componentize delete button
